formatting dates + split classes
